Add a 'monitoring' app
Use it for the definition of every model related to the monitoring of
hives.

Actually there is 3 mdoels : Hive, Captor and Data, with many-to-one
relation between them.

The view getData actually receive a post request from a hive to parse
the json and save the new object if it's correct

**TODO** : work on security by adding maybe https features, for ensuring
  the datas really come from the hive and not some bad dudes.